move redux store
