+++ pages/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-import styled from 'styled-components';
 import Counter from "../components/Counter";
 
 function Home() {
