+++ redux/store.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
 
-import reducer from "./redux/reducers";
+import reducer from "./reducers";
 
 const exampleInitialState = {
   count: 0
